Change client mode to write using the non-blocking operation (#52)
This commit changes the client mode to write using the write_nonblock method. The motivation for changing it is the current issues with IO.select and SSL, especially when using TLSv1.3.
According to the IO docs, the readability notified by IO.select doesn’t necessarily mean readability from the OpenSSL::SSL::SSLSocket object, which might lead to a blocking operation, letting the plugin in a hang state when it tries to read from the socket. Removing the reading part of the existing code would also fix this issue, but write_nonblock is the official recommendation.